Refactor graphhopper api * rename directory from 'routing' -> 'api' * Make bundle api-calls into class * inject Api class where necessary * separate api type declaration and api implementation * RouteRequestFailed action takes error message instead of response
