+++ src/map/Map.tsx
@@ -1,10 +1,10 @@
-import { Bbox, Path } from '@/routing/Api'
 import { QueryPoint } from '@/stores/QueryStore'
 import React, { useEffect, useRef, useState } from 'react'
 import styles from '@/map/Map.module.css'
 import Mapbox from '@/map/Mapbox'
 import Dispatcher from '@/stores/Dispatcher'
 import { ClearPoints, SetPoint } from '@/actions/Actions'
+import { Bbox, Path } from '@/api/graphhopper'
 
 type MapProps = {
     selectedPath: Path
