+++ src/map/Mapbox.ts
@@ -4,8 +4,8 @@ import { QueryPoint } from '@/stores/QueryStore'
 import Dispatcher from '@/stores/Dispatcher'
 import { SetPoint, SetSelectedPath } from '@/actions/Actions'
 import { Popup } from '@/map/Popup'
-import { Bbox, Path } from '@/routing/Api'
 import { FeatureCollection, LineString } from 'geojson'
+import { Bbox, Path } from '@/api/graphhopper'
 
 const selectedPathSourceKey = 'selectedPathSource'
 const selectedPathLayerKey = 'selectedPathLayer'
