+++ test/routing/Api.test.ts
@@ -1,7 +1,8 @@
 import fetchMock from 'jest-fetch-mock'
-import { ApiInfo, ErrorResponse, ghKey, info, RawResult, route, RoutingArgs, RoutingRequest } from '@/routing/Api'
 import Dispatcher, { Action } from '../../src/stores/Dispatcher'
 import { InfoReceived, RouteRequestFailed, RouteRequestSuccess } from '../../src/actions/Actions'
+import Api, { ghKey } from '../../src/api/Api'
+import { ApiInfo, ErrorResponse, RawResult, RoutingArgs, RoutingRequest } from '../../src/api/graphhopper'
 
 // replace global 'fetch' method by fetchMock
 beforeAll(fetchMock.enableMocks)
@@ -48,7 +49,7 @@ describe('info api', () => {
             },
         })
 
-        await info()
+        await new Api().infoWithDispatch()
     })
 
     it('should convert the response into an ApiInfo object', async () => {
@@ -83,7 +84,7 @@ describe('info api', () => {
             },
         })
 
-        await info()
+        await new Api().infoWithDispatch()
     })
 })
 
@@ -105,7 +106,7 @@ describe('route', () => {
             return Promise.resolve(JSON.stringify(getEmptyResult()))
         })
 
-        await route(args)
+        await new Api().routeWithDispatch(args)
     })
 
     it('transforms routingArgs into routing request with default algorithm for maxAlternativeRoutes: 1', async () => {
@@ -130,7 +131,7 @@ describe('route', () => {
             return compareRequestBodyAndResolve(request, expectedBody)
         })
 
-        await route(args)
+        await new Api().routeWithDispatch(args)
     })
 
     it('transforms routingArgs into routing request with alternative_route algorithm for maxAlternativeRoutes > 1', async () => {
@@ -157,7 +158,7 @@ describe('route', () => {
             return compareRequestBodyAndResolve(request, expectedBody)
         })
 
-        await route(args)
+        await new Api().routeWithDispatch(args)
     })
 
     it('should create an action when a response is received', async () => {
@@ -180,7 +181,7 @@ describe('route', () => {
             },
         })
 
-        await route(args)
+        await new Api().routeWithDispatch(args)
     })
 
     it('should create an action when an error is received', async () => {
@@ -203,12 +204,12 @@ describe('route', () => {
         Dispatcher.register({
             receive(action: Action) {
                 expect(action instanceof RouteRequestFailed).toBeTruthy()
-                expect((action as RouteRequestFailed).error.message).toEqual(error.message)
+                expect((action as RouteRequestFailed).errorMessage).toEqual(error.message)
                 expect((action as RouteRequestFailed).request).toEqual(args)
             },
         })
 
-        await route(args)
+        await new Api().routeWithDispatch(args)
     })
 })
